refactor move drug class into its own file
