+++ index.js
@@ -1,4 +1,5 @@
-import { Drug, Pharmacy } from "./pharmacy";
+import { Pharmacy } from "./pharmacy";
+import { Drug } from "./drug";
 
 import fs from "fs";
 
@@ -6,7 +7,7 @@ const drugs = [
   new Drug("Doliprane", 20, 30),
   new Drug("Herbal Tea", 10, 5),
   new Drug("Fervex", 5, 40),
-  new Drug("Magic Pill", 15, 40)
+  new Drug("Magic Pill", 15, 40),
 ];
 const trial = new Pharmacy(drugs);
 
@@ -17,7 +18,7 @@ for (let elapsedDays = 0; elapsedDays < 30; elapsedDays++) {
 }
 
 /* eslint-disable no-console */
-fs.writeFile("output.txt", log, err => {
+fs.writeFile("output.txt", log, (err) => {
   if (err) {
     console.log("error");
   } else {
